Moving CL_NATIVE_BOOK branch to become the trunk
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@3900 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ ExternalTools/externaltooldlg.cpp
@@ -187,7 +187,7 @@ std::vector<ToolInfo> ExternalToolDlg::GetTools()
 void ExternalToolDlg::SetTools(const std::vector<ToolInfo>& tools)
 {
 	m_listCtrlTools->Freeze();
-	for(size_t i=0; i<m_listCtrlTools->GetItemCount(); i++) {
+	for(size_t i=0; i<(size_t)m_listCtrlTools->GetItemCount(); i++) {
 		ExternalToolData* data = (ExternalToolData*)m_listCtrlTools->GetItemData(i);
 		if(data) {
 			delete data;
